settings.py
import os
from dotenv import load_dotenv

load_dotenv('.env')
url_to_db = f"postgresql+psycopg2://{os.getenv('POSTGRES_USER')}:{os.getenv('POSTGRES_PASSWORD')}@localhost:5432/{os.getenv('POSTGRES_DB')}"
