pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-06"
version = "0.1.0"
description = "University DB"
authors = ["Alex Goliansky <alx5@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.8"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
faker = "^33.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
